% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anniv.foulek.R
\name{anniv.foulek}
\alias{anniv.foulek}
\title{Pour afficher un message d'anniversaire dans R}
\usage{
anniv.foulek()
}
\value{
Crache pas du picon au plafond sale foulek
}
\description{
Pour afficher un message d'anniversaire dans R
}
\examples{
anniv.foulek()
}
